Added hilt dependency injection.
