+++ app/build.gradle
@@ -6,6 +6,10 @@ apply plugin: 'app.cash.sqldelight'
 
 apply plugin: "androidx.navigation.safeargs.kotlin"
 
+apply plugin: 'kotlin-kapt'
+
+apply plugin: 'dagger.hilt.android.plugin'
+
 
 android {
     compileSdk = 34
@@ -26,14 +30,15 @@ android {
     }
 
     viewBinding.enabled = true
+
     compileOptions {
-        targetCompatibility "1.8"
-        sourceCompatibility "1.8"
+        sourceCompatibility JavaVersion.VERSION_17
+        targetCompatibility JavaVersion.VERSION_17
+    }
+    kotlinOptions {
+        jvmTarget = '17'
     }
 
-    kotlinOptions {
-        jvmTarget = "1.8"
-    }
     namespace 'me.elmanss.melate'
 }
 
@@ -51,7 +56,6 @@ dependencies {
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.2.1'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.6.1'
-    implementation 'androidx.core:core-ktx:1.13.1'
 
     //Android Jetpack
     implementation 'androidx.navigation:navigation-fragment-ktx:2.7.7'
@@ -59,7 +63,17 @@ dependencies {
     implementation 'androidx.navigation:navigation-ui-ktx:2.7.7'
     // For Kotlin use navigation-ui-ktx
 
+
+    implementation "com.google.dagger:hilt-android:2.51.1"
+    kapt "com.google.dagger:hilt-compiler:2.51.1"
+
+
 }
+
+kapt {
+    correctErrorTypes true
+}
+
 repositories {
     google()
     mavenCentral()
