+++ app/src/androidTest/java/me/elmanss/melate/ExampleInstrumentedTest.kt
@@ -2,7 +2,7 @@ package me.elmanss.melate
 
 import androidx.test.InstrumentationRegistry
 import androidx.test.runner.AndroidJUnit4
-import org.junit.Assert.*
+import org.junit.Assert.assertEquals
 import org.junit.Test
 import org.junit.runner.RunWith
