+++ app/src/main/java/me/elmanss/melate/Melate.kt
@@ -1,29 +1,15 @@
 package me.elmanss.melate
 
 import android.app.Application
-import app.cash.sqldelight.db.SqlDriver
-import app.cash.sqldelight.driver.android.AndroidSqliteDriver
+import dagger.hilt.android.HiltAndroidApp
 import logcat.AndroidLogcatLogger
 import logcat.LogPriority
 
+@HiltAndroidApp
 class Melate : Application() {
-
-  companion object {
-    private lateinit var instance: Melate
-
-    fun get(): Melate {
-      return instance
-    }
-  }
-
-  lateinit var database: Database
 
   override fun onCreate() {
     super.onCreate()
-    val driver: SqlDriver = AndroidSqliteDriver(Database.Schema, this, "favoritos.db")
-    Database.Schema.create(driver)
-    database = Database(driver)
     AndroidLogcatLogger.installOnDebuggableApp(this, minPriority = LogPriority.VERBOSE)
-    instance = this
   }
 }
