+++ app/src/main/java/me/elmanss/melate/common/data/repository/FavoritosRepository.kt
@@ -1,0 +1,14 @@
+package me.elmanss.melate.common.data.repository
+
+import app.cash.sqldelight.Query
+import kotlinx.coroutines.flow.Flow
+import me.elmanss.melate.data.Favorito
+
+interface FavoritosRepository {
+
+  suspend fun createFavoritos(sorteoString: String)
+
+  fun selectAllFavoritos(): Flow<Query<Favorito>>
+
+  suspend fun deleteFavorito(favoritoId: Long)
+}
