+++ app/src/main/java/me/elmanss/melate/common/presentation/ui/MainActivity.kt
@@ -1,0 +1,14 @@
+package me.elmanss.melate.common.presentation.ui
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import dagger.hilt.android.AndroidEntryPoint
+import me.elmanss.melate.R
+
+@AndroidEntryPoint
+class MainActivity : AppCompatActivity() {
+  override fun onCreate(savedInstanceState: Bundle?) {
+    super.onCreate(savedInstanceState)
+    setContentView(R.layout.activity_main)
+  }
+}
