+++ app/src/main/java/me/elmanss/melate/common/util/Extensions.kt
@@ -1,0 +1,5 @@
+package me.elmanss.melate.common.util
+
+fun List<String>?.prettyPrint(): String {
+  return this?.joinToString(", ") ?: ""
+}
