+++ app/src/main/java/me/elmanss/melate/favorites/domain/model/FavoritoModel.kt
@@ -1,0 +1,3 @@
+package me.elmanss.melate.favorites.domain.model
+
+data class FavoritoModel(val id: Long, val sorteo: String)
