+++ app/src/main/java/me/elmanss/melate/favorites/domain/usecase/FavoritesUseCases.kt
@@ -1,0 +1,7 @@
+package me.elmanss.melate.favorites.domain.usecase
+
+data class FavoritesUseCases(
+  val addFavorite: AddFavorite,
+  val deleteFavorite: DeleteFavorite,
+  val fetchFavorites: FetchFavorites,
+)
