+++ app/src/main/java/me/elmanss/melate/favorites/presentation/list/ListFavoritesScreenState.kt
@@ -1,0 +1,5 @@
+package me.elmanss.melate.favorites.presentation.list
+
+import me.elmanss.melate.favorites.domain.model.FavoritoModel
+
+data class ListFavoritesScreenState(val favs: List<FavoritoModel> = emptyList())
