+++ app/src/main/java/me/elmanss/melate/home/data/datasource/remote/SorteoApi.kt
@@ -1,0 +1,5 @@
+package me.elmanss.melate.home.data.datasource.remote
+
+interface SorteoApi {
+  fun fetchSorteos(): List<Int>
+}
