+++ app/src/main/java/me/elmanss/melate/home/data/datasource/remote/SorteoApiImpl.kt
@@ -1,0 +1,70 @@
+package me.elmanss.melate.home.data.datasource.remote
+
+import java.util.concurrent.ThreadLocalRandom
+import javax.inject.Inject
+
+class SorteoApiImpl @Inject constructor(private val random: ThreadLocalRandom) : SorteoApi {
+  override fun fetchSorteos(): List<Int> {
+    return NUMBERS_POOL.shuffled(random).take(6).sorted()
+  }
+}
+
+private val NUMBERS_POOL =
+  listOf(
+    1,
+    2,
+    3,
+    4,
+    5,
+    6,
+    7,
+    8,
+    9,
+    10,
+    11,
+    12,
+    13,
+    14,
+    15,
+    16,
+    17,
+    18,
+    19,
+    20,
+    21,
+    22,
+    23,
+    24,
+    25,
+    26,
+    27,
+    28,
+    29,
+    30,
+    31,
+    32,
+    33,
+    34,
+    35,
+    36,
+    37,
+    38,
+    39,
+    40,
+    41,
+    42,
+    43,
+    44,
+    45,
+    46,
+    47,
+    48,
+    49,
+    50,
+    51,
+    52,
+    53,
+    54,
+    55,
+    56,
+  )
