+++ app/src/main/java/me/elmanss/melate/home/data/repository/SorteoRepository.kt
@@ -1,0 +1,6 @@
+package me.elmanss.melate.home.data.repository
+
+interface SorteoRepository {
+
+  fun fetchSorteos(): List<Int>
+}
