+++ app/src/main/java/me/elmanss/melate/home/domain/usecase/HomeUseCases.kt
@@ -1,0 +1,3 @@
+package me.elmanss.melate.home.domain.usecase
+
+data class HomeUseCases(val fetchSorteos: FetchSorteos, val saveToFavorites: SaveToFavorites)
