+++ app/src/main/java/me/elmanss/melate/home/presentation/HomeScreenState.kt
@@ -1,0 +1,8 @@
+package me.elmanss.melate.home.presentation
+
+import me.elmanss.melate.home.domain.model.SorteoModel
+
+data class HomeScreenState(
+  val sorteos: List<SorteoModel> = emptyList(),
+  val isWarningShown: Boolean = false,
+)
