--- app/src/main/java/me/elmanss/melate/ui/MainActivity.kt
@@ -1,12 +1,0 @@
-package me.elmanss.melate.ui
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import me.elmanss.melate.R
-
-class MainActivity : AppCompatActivity() {
-  override fun onCreate(savedInstanceState: Bundle?) {
-    super.onCreate(savedInstanceState)
-    setContentView(R.layout.activity_main)
-  }
-}
