+++ app/src/main/res/layout/fragment_add_to_fav.xml
@@ -5,7 +5,7 @@
         android:id="@+id/screen_add_to_fav"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.MainActivity">
+        tools:context=".common.presentation.ui.MainActivity">
 
     <TextView
             android:id="@+id/tv_capture_number"
