+++ app/src/main/res/layout/fragment_favs.xml
@@ -6,7 +6,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.MainActivity">
+        tools:context=".common.presentation.ui.MainActivity">
 
 
     <androidx.recyclerview.widget.RecyclerView
