+++ app/src/main/res/layout/fragment_sorteo_list.xml
@@ -5,7 +5,7 @@
         android:id="@+id/mainScreen"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.MainActivity">
+        tools:context=".common.presentation.ui.MainActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
