+++ build.gradle
@@ -1,8 +1,6 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext.kotlin_version = '1.6.20-M1'
-    ext.kotlin_version = '1.6.21'
     repositories {
         google()
         mavenCentral()
@@ -10,11 +8,12 @@ buildscript {
     }
     dependencies {
         classpath 'com.android.tools.build:gradle:8.5.1'
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:2.0.0"
         classpath 'app.cash.sqldelight:gradle-plugin:2.0.2'
 
         // Jetpack
         classpath "androidx.navigation:navigation-safe-args-gradle-plugin:2.7.7"
+        classpath 'com.google.dagger:hilt-android-gradle-plugin:2.51.1'
 
 
         // NOTE: Do not place your application dependencies here; they belong
